#131 - Fix post-send double-free in example and document TX burst ownership
send_tx_burst() already frees the burst on its NO_SPACE_AVAILABLE (full-ring)
path, so callers that free again after a failed send double-free the packet
mbufs and reuse already-returned metadata/pointer arrays.

- examples/gds_write_example.cpp: drop free_all_packets_and_burst_tx() after a
  failed send_tx_burst(); the burst is already consumed.
- include/daqiri/common.h, docs/api-reference/{cpp,python}.md: document that
  send_tx_burst() takes ownership on SUCCESS and on NO_SPACE_AVAILABLE, so the
  application must not free or access the burst afterwards. Scoped to those
  paths (INVALID_PARAMETER does not consume — see #164).
- src/managers/dpdk/daqiri_dpdk_mgr.cpp: use size_t for the inner packet loop
  counter in free_unsubmitted_tx_packets() to match num_pkts.

+++ src/engines/dpdk/daqiri_dpdk_engine.cpp
@@ -91,7 +91,7 @@ static void free_unsubmitted_tx_packets(BurstParams* burst) {
     if (burst->pkts[seg] == nullptr) { continue; }
 
     auto** pkts = reinterpret_cast<rte_mbuf**>(burst->pkts[seg]);
-    for (int pkt = 0; pkt < burst->hdr.hdr.num_pkts; pkt++) {
+    for (size_t pkt = 0; pkt < burst->hdr.hdr.num_pkts; pkt++) {
       if (pkts[pkt] != nullptr) { rte_pktmbuf_free_seg(pkts[pkt]); }
     }
   }
